// src/helpers/enums/generalEnum.ts
export enum GeneralEnum {
    Activo = "A",
    Inactivo = "I",
    Ascending = "ASC",
    Descending = "DESC",
}

export enum TipoFileTorneoEnum {
    Reglamento = "R",
    Imagen = "I",
}

export enum ImagenTypeEnum {
    base64 = "base64",
    asFile = "file",
    contentType = "image/jpeg",
}

export enum ImagenDefaultEnum {
    torneo = "portada.jpg",
    delegacion = "team.png",
    competidor = "icon_tkd.png",
}

export enum PdfTypeEnum {
    contentType = "application/pdf",
    sizeA4 = "A4",
    layoutPortrait = "portrait",
    layoutLandscape = "landscape",
    marginDefault = 20,
    textColorMain = "#0B3E60",
    textColorBlack = "#000000",
    textColorWhite = "#FFFFFF",
    textFont = "Helvetica",
    textFontBold = "Helvetica-Bold",
    textFontBoldItalic = "Helvetica-BoldOblique",
    textAlignmentCenter = "center",
    textAlignmentRight = "right",
    textAlignmentLeft = "left",
    fontSizeDefault = 10,
    fontSizeSmall = 6,
    fontSizeMedium = 12,
    fontSizeLarge = 16,
}

export enum EncriptEnum {
    Encriptar = "E",
    Desencriptar = "D",
}